app/annotation/pyproject.toml
[tool.poetry]
name = "annotation"
version = "0.1.0"
description = ""
authors = ["kohei3110 <kk.31108424@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
logging = "^0.4.9.6"
pydantic = "^2.8.2"
openai = "^1.36.1"
tenacity = "^8.5.0"
azure-core = "^1.30.2"
wikipedia = "^1.4.0"
flask = "^3.0.3"
azure-cosmos = "^4.7.0"
azure-storage-blob = "^12.21.0"
azure-identity = "^1.17.1"
uvicorn = "^0.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
